Basic Order filling added
Robot can now pickup and travel to the needed dropoff point to fill orders, not perfect by any means, the robot moves randomly when not path finding and just picks up any items to see if it's needed. It then burns the item if it isn't instead of not picking it up.